refactor(charging-station): extract serial-number/id helpers to HelpersId (#1936 f-4)
Extracts 4 identity helpers (getChargingStationId, getHashId, createSerialNumber,

propagateSerialNumber) plus the ChargingStationNameTemplate type alias into a

dedicated HelpersId.ts. The getRandomSerialNumberSuffix helper had a single

internal caller (createSerialNumber) and no external consumer; it is demoted to

module-private in the new file, matching its actual usage.

Helpers.ts continues to re-export the 4 public symbols and the type alias via

the barrel, so all 5 consumer files work unchanged.

Refs #1936
